+++ src/store/reducers.ts
@@ -3,6 +3,7 @@ import { ReducersMapObject } from 'redux'
 import app, { AppActions } from 'Store/ducks/app'
 import chatters, { ChattersActions } from 'Store/ducks/chatters'
 import logs, { LogsActions } from 'Store/ducks/logs'
+import notes, { NotesActions } from 'Store/ducks/notes'
 import settings, { SettingsActions } from 'Store/ducks/settings'
 import user, { UserActions } from 'Store/ducks/user'
 
@@ -16,6 +17,7 @@ const reducers: ReducersMapObject<ApplicationState, ApplicationActions | any> = 
   app,
   chatters,
   logs,
+  notes,
   settings,
   user,
 }
@@ -25,7 +27,7 @@ export default reducers
 /**
  * Application actions
  */
-type ApplicationActions = AppActions | LogsActions | UserActions | ChattersActions | SettingsActions
+type ApplicationActions = AppActions | LogsActions | NotesActions | UserActions | ChattersActions | SettingsActions
 
 /**
  * Application state.
@@ -34,6 +36,7 @@ export interface ApplicationState {
   app: ReturnType<typeof app>
   chatters: ReturnType<typeof chatters>
   logs: ReturnType<typeof logs>
+  notes: ReturnType<typeof notes>
   user: ReturnType<typeof user>
   settings: ReturnType<typeof settings>
 }